Address PR review: move current-semester tests to unit suite
